Conflicts:
	.project

+++ src/main/java/com/pse/testserver/entities/Message.java
@@ -3,28 +3,53 @@ package com.pse.testserver.entities;
 import javax.persistence.*;
 import java.util.Date;
 
+/**
+ * Entity class for the persisted message data from the "messages" table.
+ * Represents a message, providing access to the message's unique id, text,
+ * sender, receiver and date of creation
+ */
 @Entity
 @Table(name = "messages")
 public class Message {
 
+    /**
+     * Incremental generated unique id.
+     */
     @Column(name = "id")
-    private int id;
+    private long id;
 
+    /**
+     * User-created text to be displayed in this message.
+     */
     @Column(name = "text")
     private String text;
 
+    /**
+     * User, which sent this message.
+     */
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "sender_id")
     private User sender;
 
+    /**
+     * User, which received this message.
+     */
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "receiver_id")
     private User receiver;
 
+    /**
+     * Date of creation of this message.
+     */
     @Column(name = "date")
     private Date date;
 
-    public int getId() {
+    /**
+     * Gets id.
+     *
+     * @return Value of id.
+     */
+    public long getId() {
         return id;
     }
